Update custom.js
We are working with H5P to address a lack of display issue for iframe embedded H5P content in summary tags. I've added the code the vendor has sent and will test it on the test server once it has been put in place. I'll also put in a ticket for this. Thanks! Nancy