Update mobile.css
I had two classes where the text color was relying on something that was changed so I've added a color to the .gbanner info (line 213) and the instructorBox class to a group of similar classes (line 728ish).  I also changed the highlight color to something that works with both black and white text (line 467).